:sparkles: Create dedicated package for workflow integration testing
