+++ src/platform/packages/shared/kbn-connector-specs-test-helpers/kibana.jsonc
@@ -1,0 +1,10 @@
+{
+  "type": "test-helper",
+  "id": "@kbn/connector-specs-test-helpers",
+  "owner": [
+    "@elastic/workchat-eng"
+  ],
+  "group": "platform",
+  "visibility": "shared",
+  "devOnly": true
+}
