+++ tsconfig.base.json
@@ -272,6 +272,8 @@
       "@kbn/connector-schemas/*": ["src/platform/packages/shared/kbn-connector-schemas/*"],
       "@kbn/connector-specs": ["src/platform/packages/shared/kbn-connector-specs"],
       "@kbn/connector-specs/*": ["src/platform/packages/shared/kbn-connector-specs/*"],
+      "@kbn/connector-specs-test-helpers": ["src/platform/packages/shared/kbn-connector-specs-test-helpers"],
+      "@kbn/connector-specs-test-helpers/*": ["src/platform/packages/shared/kbn-connector-specs-test-helpers/*"],
       "@kbn/console-plugin": ["src/platform/plugins/shared/console"],
       "@kbn/console-plugin/*": ["src/platform/plugins/shared/console/*"],
       "@kbn/content-connectors-plugin": ["x-pack/platform/plugins/shared/content_connectors"],
